+++ test_resolwe_runtime_utils.py
@@ -18,8 +18,8 @@ from unittest import TestCase
 from unittest.mock import patch
 
 from resolwe_runtime_utils import (
-        save, save_list, save_file, save_file_list, error, warning, info, progress, checkrc
-    )
+    save, save_list, save_file, save_dir, save_file_list,
+    save_dir_list, error, warning, info, progress, checkrc)
 
 
 class ResolweRuntimeUtilsTestCase(TestCase):
@@ -131,6 +131,59 @@ class TestGenSaveFileList(ResolweRuntimeUtilsTestCase):
             save_file_list('src', 'foo.py:ref1.gz:ref2.gz', 'bar.py'),
             '{"proc.error": "Only one colon \':\' allowed in file-refs specification."}'
         )
+
+
+class TestGenSaveDir(ResolweRuntimeUtilsTestCase):
+
+    @patch('os.path.isdir', return_value=True)
+    @patch('os.path.isfile', return_value=True)
+    def test_file(self, isfile_mock, isdir_mock):
+        self.assertEqual(save_dir('etc', 'foo'),
+                         '{"etc": {"dir": "foo"}}')
+        self.assertEqual(save_dir('etc', 'foo bar'),
+                         '{"etc": {"dir": "foo bar"}}')
+
+    def test_file_missing(self):
+        self.assertEqual(save_dir('etc', 'foo'),
+                         '{"proc.error": "Output \'etc\' set to a missing directory: \'foo\'."}')
+        self.assertEqual(save_dir('etc', 'foo bar'),
+                         '{"proc.error": "Output \'etc\' set to a missing directory: \'foo bar\'."}')
+
+    @patch('os.path.isdir', return_value=True)
+    @patch('os.path.isfile', return_value=True)
+    def test_file_with_refs(self, isfile_mock, isdir_mock):
+        self.assertJSONEqual(
+            save_dir('etc', 'foo', 'ref1.txt', 'ref2.txt'),
+            '{"etc": {"dir": "foo", "refs": ["ref1.txt", "ref2.txt"]}}'
+        )
+
+    def test_improper_input(self):
+        self.assertRaises(TypeError, save_dir, 'etc')
+
+
+class TestGenSaveDirList(ResolweRuntimeUtilsTestCase):
+
+    @patch('os.path.isdir', return_value=True)
+    @patch('os.path.isfile', return_value=True)
+    def test_files(self, isfile_mock, isdir_mock):
+        self.assertEqual(save_dir_list('src', 'dir1', 'dir 2', 'dir/3'),
+            '{"src": [{"dir": "dir1"}, {"dir": "dir 2"}, {"dir": "dir/3"}]}')
+
+    def test_files_missing(self):
+        self.assertEqual(save_dir_list('src', 'dir1', 'dir 2', 'dir/3'),
+            '{"proc.error": "Output \'src\' set to a missing directory: \'dir1\'."}')
+
+    @patch('os.path.isdir', return_value=True)
+    @patch('os.path.isfile', return_value=True)
+    def test_files_with_refs(self, isfile_mock, isdir_mock):
+        self.assertEqual(json.loads(save_dir_list('src', 'dir1:ref1.bar,ref2.bar,ref3.bar', 'dir2')),
+        {"src": [{"refs": ["ref1.bar", "ref2.bar", "ref3.bar"], "dir": "dir1"}, {"dir": "dir2"}]})
+
+    @patch('os.path.isdir', return_value=True)
+    @patch('os.path.isfile', return_value=True)
+    def test_files_invalid_format(self, isfile_mock, isdir_mock):
+        self.assertEqual(save_dir_list('src', 'dir1:ref1.bar:ref2.bar', 'dir2'),
+        '{"proc.error": "Only one colon \':\' allowed in directory."}')
 
 
 class TestGenInfo(ResolweRuntimeUtilsTestCase):